services/api-gateway/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Monitoring
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Logging
structlog>=24.1.0

# Utils
python-dotenv>=1.0.0
